+++ tests/test_completers.py
@@ -1,10 +1,15 @@
+import time
+import threading
 from unittest.mock import MagicMock
 from prompt_toolkit.completion import Completion
 from prompt_toolkit.formatted_text import FormattedText
 
 from iredis.completers import LatestUsedFirstWordCompleter
 from iredis.config import config
+from iredis.entry import compile_grammar_bg
 from .conftest import prompt_session
+
+
 
 
 def test_LUF_completer_touch():
@@ -85,3 +90,19 @@ def test_newbie_mode_complete_with_meta_dict():
             ),
         ),
     ]
+
+
+def test_patch_grammer_and_session_after_startup():
+    session = MagicMock()
+    normal_thread_count = threading.active_count()
+    session.lexer = None
+    session.completer = None
+    compile_grammar_bg(session)
+    assert session.lexer is None
+    assert session.completer is None
+    assert threading.active_count() == normal_thread_count + 1
+
+    while threading.active_count() > normal_thread_count:
+        time.sleep(0.1)
+    assert session.lexer
+    assert session.completer
